CIS Octuparra and CIS Artillery added
Yo man, both land units have been coded but there are some issues in
game:

The Octuparra fire cones need to be reduced, the lasers can fire at a 90
angle degree... looks unrealistic. Also the deploy and undeploy
animations break when the unit is firing, can you please have a look?

The CIS Artillery is really giving me a headache, the projectile fires,
but it does no damage whatsoever to any units (Republic or Tusken), this
also made me realize that the tusken raiders in tattooine are not being
hostile to the CIS or Republic (perhaps this is a gameconstans issue?)

Please have a look and let me know man.