Shocktroopers, Evo, and Mando clones modifications (version 2)
Alright to I gave these units the Remote Bomb ability instead of sticky
bomb, I changed the <supports_autofire> tag to false in all of them.

The EVO troopers don't have an ability anymore as their throw grenade
animation is not 100% there. This seems to have fixed the crash to
desktop along with changing the "Saboteur_remote_grenade" definition to
"Shocktrooper_remote_grenade"...  As the code seemed to search for a
Saboteur in the field but the unit wasn't coded as an infiltrator in the
first place.

Please see what you think and let me know what solution you think is
best, I'd say let's go back to the Remote grenade as this ability has a
detonation time. Alternatively we could give the EVO the sticky bomb
ability.

This might conflict with your changes, this is why I'm creating a branch
so this particular fix remains separate from the master until we decide
what to do.

There are also some minor fixes to scale factors, fire pulses and
changed the predator SD ability icon to something more accurate instead
of the s foil.