Audio and rebel voice and fixes
So here is all the audio so far and some code fixes along with the rebel
sniper voices.

When you go to the cause in make sure you are replacing it in the same
SFXevent and not making new ones, also you need to include the file
patch to said audio so it'd be like this
Data\Audio\SFX\Battlefront\Stormtooper\thenaudioname.wav

You will aslo need to add this under the Samples or it will be super
quiet in game.

<Min_Volume>80</Min_Volume>
<Max_Volume>80</Max_Volume>