TCW Era balancing tweaks and improvements
Please have  a look at the files - I tweaked several things in regards
to spawned squadrons and tech levels of construction, e.g. the Hyena
bombers will not be available until tech 2, same for the V wings for the
republic.

Also the Sensor arrays now have a decent FOW reveal range.

I'll code in the BX battledroid and the Porax fighter for the CIS next.

Happy halloween!