Many Fixes
So in this i fixed a few thing added plo-koon,kit-fisto and edited the
jedi along with a few other major changes and the jedi temple is now
buildable.

I may have overwrote the commander fox icon so please send it over skype
the next time we talk not the MTD files but the .tga file for the icon
itself so i can add it and you'll as well wanna add it back.